Camping-main/WebCamping/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
asgiref==3.8.1
    # via
    #   django
    #   django-cors-headers
certifi==2024.2.2
    # via requests
charset-normalizer==3.3.2
    # via requests
django==5.0.4
    # via
    #   -r requirements.in
    #   django-admin-honeypot-updated-2021
    #   django-axes
    #   django-cors-headers
    #   django-filter
    #   djangorestframework
django-admin-honeypot-updated-2021==1.2.0
    # via -r requirements.in
django-axes==6.4.0
    # via -r requirements.in
django-cors-headers==4.3.1
    # via -r requirements.in
django-cors-middleware==1.5.0
    # via -r requirements.in
django-environ==0.11.2
    # via -r requirements.in
django-filter==24.2
    # via -r requirements.in
django-ipware==6.0.5
    # via django-admin-honeypot-updated-2021
djangorestframework==3.15.1
    # via -r requirements.in
googlemaps==4.10.0
    # via -r requirements.in
idna==3.7
    # via requests
markdown==3.6
    # via -r requirements.in
psycopg2==2.9.9
    # via -r requirements.in
python-ipware==2.0.4
    # via django-ipware
requests==2.31.0
    # via googlemaps
sqlparse==0.5.0
    # via django
tzdata==2024.1
    # via django
urllib3==2.2.1
    # via requests
